Extract LoopDominanceReportSection from ResultReportGenerator (#1444)
Move loop dominance report generation (stacked area chart SVG,
dominance transition table, loop color logic) into a dedicated
LoopDominanceReportSection class. ResultReportGenerator delegates
to the section class and retains shared utilities.

ResultReportGenerator drops from 1,389 to 1,185 lines. Shared
constants and utilities made package-private for section class access.
